feat(ui,main-site): move community page to under core layout, create leftsidebar component
