+++ apps/main-site/src/app/(main-site)/(app)/c/[communityId]/[channelId]/page.tsx
