+++ apps/main-site/src/app/(main-site)/(app)/c/[communityId]/page.tsx
